requirements.txt
# Flask核心
Flask==3.0.0
Flask-CORS==4.0.0

# 数据处理
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2        # Excel文件读写
xlrd==2.0.1            # 旧版Excel支持

# 可视化
plotly==5.18.0

# 配置管理
python-dotenv==1.0.0

# 日志（Flask已包含logging）
# 如需更强大的日志可选：
# loguru==0.7.2

# 工具库
tqdm==4.66.1           # 进度条

# 可选：HTML转图片（如果需要）
# playwright==1.40.0

# 生产环境推荐（可选）
# gunicorn==21.2.0      # 生产级WSGI服务器
# gevent==23.9.1        # 异步支持
